Enhance database migration and development workflow
- Improve migration script with robust error handling and logging
- Add dynamic migration file detection and processing
- Update Docker Compose and development scripts to support migration workflow
- Modify database Dockerfile to include migration watcher
- Refactor migration error handling to continue processing on individual migration failures
- Add fallback logo rendering in frontend quote component

+++ backend/jobs/securities.go
@@ -231,7 +231,7 @@ func updateSecurityCik(conn *utils.Conn) error {
 
 	// Parse the JSON response
 	var secData map[string]struct {
-		CikStr int    `json:"cik_str"`
+		CikStr int64  `json:"cik_str"`
 		Ticker string `json:"ticker"`
 		Title  string `json:"title"`
 	}
